controllers added right place
