+++ src/main/java/com/example/wroom/controllers/CostumerController.java
@@ -1,4 +1,4 @@
-package controllers;
+package com.example.wroom.controllers;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -7,6 +7,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
  * @author:Marek Uibo
  */
 @Controller
-@RequestMapping("/rentalOffice")
-public class RentalOfficeController {
+@RequestMapping("/costumer")
+public class CostumerController {
 }
