+++ src/main/java/com/example/wroom/controllers/EmployeeController.java
@@ -1,4 +1,4 @@
-package controllers;
+package com.example.wroom.controllers;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
